backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
python-chess==1.999
python-dotenv==1.0.1
